chore: improve unit tests for XIVAPI functionality and refactor existing tests
- Seperated test files for methods used by xivapi.
- Updated TypeScript configuration to target ESNext for improved compatibility.
- Moved Vitest timeout to the global configuration to accommodate longer-running tests.
