+++ .github/dependabot.yml
@@ -12,7 +12,8 @@ updates:
     open-pull-requests-limit: 1
     ignore:
       - dependency-name: "*"
-        update-types: ["version-update:semver-patch", "version-update:semver-minor"]
+        update-types:
+          ["version-update:semver-patch", "version-update:semver-minor"]
     commit-message:
       prefix: "[skip ci]"
       include: "scope"
