+++ .vscode/extensions.json
@@ -1,0 +1,8 @@
+{
+  "recommendations": [
+    "ryanluker.vscode-coverage-gutters",
+    "dbaeumer.vscode-eslint",
+    "esbenp.prettier-vscode",
+    "rvest.vs-code-prettier-eslint"
+  ]
+}
