+++ package.json
@@ -22,6 +22,7 @@
     "eslint": "^10.5.0",
     "eslint-config-prettier": "^10.1.8",
     "globals": "^17.6.0",
+    "jiti": "^2.7.0",
     "prettier": "3.8.4",
     "ts-node": "^10.9.2",
     "tsup": "^8.5.1",
