+++ src/lib/versions.ts
@@ -1,4 +1,4 @@
-import type { Models } from "../index.js";
+import * as Models from "../models.js";
 import { CustomError, request } from "../utils.js";
 
 /**
@@ -8,10 +8,12 @@ import { CustomError, request } from "../utils.js";
 export class Versions {
   /**
    * List versions understood by the API.
+   * @returns {Promise<Models.VersionsResponse>}
    * @see https://v2.xivapi.com/api/docs#tag/versions/get/version
    */
   async all(): Promise<Models.VersionsResponse> {
     const { data, errors } = await request({ path: "/version", params: {} });
+    /* v8 ignore if -- @preserve */
     if (errors) throw new CustomError(errors[0].message);
     return data as Models.VersionsResponse;
   }
