+++ vitest.config.ts
@@ -6,5 +6,6 @@ export default defineConfig({
       provider: "v8",
       reporter: ["text", "json-summary", "lcov", "cobertura"],
     },
+    testTimeout: 10000,
   },
 });
